requirements.txt
torch==1.13.1+cu117
torchvision==0.14.1+cu117
torchaudio==0.13.1+cu117
numpy==1.23.0
matplotlib
gdown
tqdm
clean-fid
einops
Pillow
imageio
tensorboard
pandas
scikit-learn